+++ Thunderstore/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+### v1.0.2
+
+- Fix an issue where map hazards doesn't despawn
+- Update the vanilla enemies check with v55 monsters
+
 ### v1.0.1
 
-- Fix an issue where StaticNetcodeLib was an hard dependency (you don't need it)
+- Fix an issue where StaticNetcodeLib was a hard dependency (you don't need it)
 
 ### v1.0.0
 - Release version